Backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib
bcrypt==4.0.1
sqlmodel
requests
python-multipart
faiss-cpu
python-docx
pydantic
aiofiles
